set Accept header to */* if raw param is set
The (undocumented) raw parameter says to me, "I want to get arbitrary non-JSON data from this URL; please wrap it for me." However if we explicitly set the Accept header to 'application/json' we might get a 406 back from the external service. If the user explicitly set the raw param, it makes sense to accept everything, right?